# swagger/server/src/IO.Swagger/Controllers/TaskApi.cs
/*
 * ToDo and Tracker
 *
 * For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * OpenAPI spec version: 0.0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TaskApiController : ControllerBase
    { 
        /// <summary>
        /// Add a new task
        /// </summary>
        /// <param name="body"></param>
        /// <param name="apiKey"></param>
        /// <response code="201">successfully create the task</response>
        /// <response code="400">fail to create the task</response>
        /// <response code="401">API key is missing or invalid</response>
        [HttpPost]
        [Route("/api/beta/tasks")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("AddTask")]
        [SwaggerResponse(statusCode: 201, type: typeof(Task), description: "successfully create the task")]
        public virtual IActionResult AddTask([FromBody]Task body, [FromHeader]string apiKey)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Task));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            string exampleJson = null;
            exampleJson = "{\n  \"recurRule\" : \"recurRule\",\n  \"expectedStartTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"expectedTimeDuration\" : \"P1D\",\n  \"name\" : \"name\",\n  \"stepNum\" : 5,\n  \"id\" : 6,\n  \"projectIds\" : [ 1, 1 ],\n  \"state\" : \"queued\",\n  \"priority\" : \"notAnalysed\",\n  \"expectedFinishTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"assignedUser\" : 7,\n  \"events\" : [ {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  }, {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Task>(exampleJson)
                        : default(Task);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Deletes a task
        /// </summary>
        /// <param name="tid">Task id to delete</param>
        /// <param name="apiKey"></param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="401">API key is missing or invalid</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpDelete]
        [Route("/api/beta/tasks/{tid}")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("DeleteTask")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult DeleteTask([FromRoute][Required]long? tid, [FromHeader]string apiKey)
        { 
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));

            throw new NotImplementedException();
        }

        /// <summary>
        /// Find task by ID
        /// </summary>
        /// <remarks>Returns a single task</remarks>
        /// <param name="tid">ID of task to return</param>
        /// <param name="apiKey"></param>
        /// <response code="200">successful operation</response>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="401">API key is missing or invalid</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpGet]
        [Route("/api/beta/tasks/{tid}")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetTaskById")]
        [SwaggerResponse(statusCode: 200, type: typeof(Task), description: "successful operation")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult GetTaskById([FromRoute][Required]long? tid, [FromHeader]string apiKey)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Task));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));
            string exampleJson = null;
            exampleJson = "{\n  \"recurRule\" : \"recurRule\",\n  \"expectedStartTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"expectedTimeDuration\" : \"P1D\",\n  \"name\" : \"name\",\n  \"stepNum\" : 5,\n  \"id\" : 6,\n  \"projectIds\" : [ 1, 1 ],\n  \"state\" : \"queued\",\n  \"priority\" : \"notAnalysed\",\n  \"expectedFinishTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"assignedUser\" : 7,\n  \"events\" : [ {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  }, {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Task>(exampleJson)
                        : default(Task);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get all the tasks the user involves in
        /// </summary>
        /// <param name="apiKey"></param>
        /// <response code="200">successful operation</response>
        /// <response code="401">API key is missing or invalid</response>
        [HttpGet]
        [Route("/api/beta/tasks")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetTasks")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Task>), description: "successful operation")]
        public virtual IActionResult GetTasks([FromHeader]string apiKey)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Task>));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            string exampleJson = null;
            exampleJson = "[ {\n  \"recurRule\" : \"recurRule\",\n  \"expectedStartTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"expectedTimeDuration\" : \"P1D\",\n  \"name\" : \"name\",\n  \"stepNum\" : 5,\n  \"id\" : 6,\n  \"projectIds\" : [ 1, 1 ],\n  \"state\" : \"queued\",\n  \"priority\" : \"notAnalysed\",\n  \"expectedFinishTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"assignedUser\" : 7,\n  \"events\" : [ {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  }, {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  } ]\n}, {\n  \"recurRule\" : \"recurRule\",\n  \"expectedStartTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"expectedTimeDuration\" : \"P1D\",\n  \"name\" : \"name\",\n  \"stepNum\" : 5,\n  \"id\" : 6,\n  \"projectIds\" : [ 1, 1 ],\n  \"state\" : \"queued\",\n  \"priority\" : \"notAnalysed\",\n  \"expectedFinishTime\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"assignedUser\" : 7,\n  \"events\" : [ {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  }, {\n    \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"id\" : 5,\n    \"endTime\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"taskId\" : 2\n  } ]\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Task>>(exampleJson)
                        : default(List<Task>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Updates a task with form data. Can be used to add an event
        /// </summary>
        /// <param name="body"></param>
        /// <param name="tid">ID of the task that needs to be updated</param>
        /// <param name="apiKey"></param>
        /// <response code="204">successful operation</response>
        /// <response code="400">Invalid input</response>
        /// <response code="401">API key is missing or invalid</response>
        /// <response code="404">The specified resource was not found</response>
        [HttpPut]
        [Route("/api/beta/tasks/{tid}")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("UpdateTask")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "The specified resource was not found")]
        public virtual IActionResult UpdateTask([FromBody]Task body, [FromRoute][Required]long? tid, [FromHeader]string apiKey)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(Error));

            throw new NotImplementedException();
        }
    }
}