# swagger/server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ToDo and Tracker",
    "description" : "For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version" : "0.0.1"
  },
...
  "servers" : [ {
    "url" : "https://zhenying.io/api/beta"
  } ],
...
  "paths" : {
...
    "/tasks" : {
      "get" : {
        "tags" : [ "task" ],
        "summary" : "Get all the tasks the user involves in",
        "operationId" : "getTasks",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Task"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "401" : {
            "description" : "API key is missing or invalid",
            "headers" : {
              "WWW_Authenticate" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "todotracker_auth" : [ "read:projects" ]
        }, {
          "api_key" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "task" ],
        "summary" : "Add a new task",
        "operationId" : "addTask",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Task"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "successfully create the task",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Task"
                }
              }
            }
          },
          "400" : {
            "description" : "fail to create the task",
            "content" : { }
          },
          "401" : {
            "description" : "API key is missing or invalid",
            "headers" : {
              "WWW_Authenticate" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "todotracker_auth" : [ "write:projects", "read:projects" ]
        }, {
          "api_key" : [ ]
        } ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/tasks/{tid}" : {
      "get" : {
        "tags" : [ "task" ],
        "summary" : "Find task by ID",
        "description" : "Returns a single task",
        "operationId" : "getTaskById",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tid",
          "in" : "path",
          "description" : "ID of task to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Task"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID supplied",
            "content" : { }
          },
          "401" : {
            "description" : "API key is missing or invalid",
            "headers" : {
              "WWW_Authenticate" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "todotracker_auth" : [ "read:projects" ]
        }, {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "task" ],
        "summary" : "Updates a task with form data. Can be used to add an event",
        "operationId" : "updateTask",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tid",
          "in" : "path",
          "description" : "ID of the task that needs to be updated",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Task"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "successful operation",
            "content" : { }
          },
          "400" : {
            "description" : "Invalid input",
            "content" : { }
          },
          "401" : {
            "description" : "API key is missing or invalid",
            "headers" : {
              "WWW_Authenticate" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "todotracker_auth" : [ "write:projects", "read:projects" ]
        }, {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "task" ],
        "summary" : "Deletes a task",
        "operationId" : "deleteTask",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tid",
          "in" : "path",
          "description" : "Task id to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied",
            "content" : { }
          },
          "401" : {
            "description" : "API key is missing or invalid",
            "headers" : {
              "WWW_Authenticate" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "todotracker_auth" : [ "write:projects", "read:projects" ]
        }, {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Task" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "projectIds" : {
            "type" : "array",
            "xml" : {
              "name" : "projectId",
              "wrapped" : true
            },
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "notAnalysed", "high", "medium", "low" ]
          },
          "stepNum" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedStartTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedFinishTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedTimeDuration" : {
            "type" : "string",
            "description" : "[ISO_8601](https://en.wikipedia.org/wiki/ISO_8601#Durations)",
            "format" : "iso8601",
            "example" : "P1D"
          },
          "state" : {
            "type" : "string",
            "enum" : [ "queued", "started", "blocked", "completed", "abandoned" ]
          },
          "recurRule" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "xml" : {
              "name" : "event",
              "wrapped" : true
            },
            "items" : {
              "$ref" : "#/components/schemas/Event"
            }
          },
          "assignedUser" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "example" : {
          "recurRule" : "recurRule",
          "expectedStartTime" : "2000-01-23T04:56:07.000+00:00",
          "expectedTimeDuration" : "P1D",
          "name" : "name",
          "stepNum" : 5,
          "id" : 6,
          "projectIds" : [ 1, 1 ],
          "state" : "queued",
          "priority" : "notAnalysed",
          "expectedFinishTime" : "2000-01-23T04:56:07.000+00:00",
          "assignedUser" : 7,
          "events" : [ {
            "startTime" : "2000-01-23T04:56:07.000+00:00",
            "id" : 5,
            "endTime" : "2000-01-23T04:56:07.000+00:00",
            "taskId" : 2
          }, {
            "startTime" : "2000-01-23T04:56:07.000+00:00",
            "id" : 5,
            "endTime" : "2000-01-23T04:56:07.000+00:00",
            "taskId" : 2
          } ]
        }
      },
      "Event" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taskId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "example" : {
          "startTime" : "2000-01-23T04:56:07.000+00:00",
          "id" : 5,
          "endTime" : "2000-01-23T04:56:07.000+00:00",
          "taskId" : 2
        }
      },
...
      "Error" : {
        "required" : [ "code", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
    },
...
    "securitySchemes" : {
      "todotracker_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://todotracker.zhenying.io/oauth/dialog",
            "scopes" : {
              "write:projects" : "modify projects",
              "read:projects" : "read projects"
            }
          }
        }
      },
      "petstore_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://petstore.zhenying.io/oauth/dialog",
            "scopes" : {
              "write:pets" : "modify pets in your account",
              "read:pets" : "read your pets"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}